refactor: implement application feature
