+++ src/lib/consts/routes.ts
@@ -1,3 +1,11 @@
+export const ApplicationRoutes = {
+  GENERAL_ROUTE: '/user/application/general',
+  PRIORITIZE_ROUTE: '/user/application/prioritize',
+  SELECT_ROUTE: '/user/application/select',
+  COVER_LETTERS_ROUTE: '/user/application/cover-letters',
+  REVIEW_ROUTE: '/user/application/review',
+} as const;
+
 export const AdminRoutes = {
   DASHBOARD: '/admin',
   EVENTS: '/admin/event',
